Build: Update build system and Python requirements
Update build configuration and dependencies to support OneCrypto development.

Changes:
- pip-requirements.txt: Use ~= for compatible version ranges
  * edk2-pytool-library~=0.23.10
  * edk2-pytool-extensions~=0.30.5
- Remove *FlavorBuild.py scripts in favor of a unified PlatformBuild.py

These changes ensure consistent build environments and enable building
the new OneCryptoPkg and CryptoBinPkg packages.
