+++ pip-requirements.txt
@@ -12,8 +12,8 @@
 # https://www.python.org/dev/peps/pep-0440/#version-specifiers
 ##
 
-edk2-pytool-library==0.19.8
-edk2-pytool-extensions==0.26.4
+edk2-pytool-library~=0.23.10
+edk2-pytool-extensions~=0.30.5
 regex
 GitPython
 setuptools
